Removed the multiplier for executors because it's not really necessary. This way we can set the executors in the slave charm and be reflected in the master charm.
Changed the way master-relation-departed handles slave host name.
Jenkins doesn't allow illegal characters (e.g. /) in the naming of
slave hosts when adding them, so the workaround was to replace
"/" in the juju slave host with "-"
(e.g. jenkins-slave/0 => jenkins-slave-0) when sending the slaves
host through the relation. The problem was that when the relation
was removed, when deleting the node from jenkins, jenkins charm
was taking the slaves hostname from JUJU_REMOTE_UNIT that is different
from the one sent on the relation.
With this fix, the slavename is taken from the relation with the
jenkins slave.